Rename gm_args mod to args, as it contains gm and gtf args
